+++ oral_exam2/S32_PredatorPrey/src/Deer.java
@@ -1,5 +1,5 @@
 public class Deer extends Animal {
-    public Deer(int initialSize, float growthCoefficient, float deathCoefficient) {
+    public Deer(double initialSize, double growthCoefficient, double deathCoefficient) {
         super(initialSize, growthCoefficient, deathCoefficient);
         setName("Deer");
     }
