+++ oral_exam2/S32_PredatorPrey/src/Rabbit.java
@@ -1,5 +1,5 @@
 public class Rabbit extends Animal {
-    public Rabbit(int initialSize, float growthCoefficient, float deathCoefficient) {
+    public Rabbit(double initialSize, double growthCoefficient, double deathCoefficient) {
         super(initialSize, growthCoefficient, deathCoefficient);
         setName("Rabbit");
     }
